# New: allow rules to listen for AST selectors (fixes #5407)