--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -757,6 +757,8 @@ module.exports = (function() {
      */
     api.verify = function(textOrSourceCode, config, filenameOrOptions, saveState) {
         const text = (typeof textOrSourceCode === "string") ? textOrSourceCode : null;
+        const selectors = new Set();
+
         let ast,
             parseResult,
             shebang,
@@ -868,10 +870,11 @@ module.exports = (function() {
                         : ruleCreator(ruleContext);
 
                     // add all the node types as listeners
-                    Object.keys(rule).forEach(nodeType => {
-                        api.on(nodeType, timing.enabled
-                            ? timing.time(key, rule[nodeType])
-                            : rule[nodeType]
+                    Object.keys(rule).forEach(selector => {
+                        selectors.add(selector);
+                        api.on(selector, timing.enabled
+                            ? timing.time(key, rule[selector])
+                            : rule[selector]
                         );
                     });
                 } catch (ex) {
@@ -911,7 +914,7 @@ module.exports = (function() {
                 }
             }
 
-            let eventGenerator = new NodeEventGenerator(api);
+            let eventGenerator = new NodeEventGenerator(api, Array.from(selectors));
 
             eventGenerator = new CodePathAnalyzer(eventGenerator);
             eventGenerator = new CommentEventGenerator(eventGenerator, sourceCode);