--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -914,7 +914,7 @@ module.exports = (function() {
                 }
             }
 
-            let eventGenerator = new NodeEventGenerator(api, Array.from(selectors));
+            let eventGenerator = new NodeEventGenerator(api, selectors);
 
             eventGenerator = new CodePathAnalyzer(eventGenerator);
             eventGenerator = new CommentEventGenerator(eventGenerator, sourceCode);