--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -757,8 +757,6 @@ module.exports = (function() {
      */
     api.verify = function(textOrSourceCode, config, filenameOrOptions, saveState) {
         const text = (typeof textOrSourceCode === "string") ? textOrSourceCode : null;
-        const selectors = new Set();
-
         let ast,
             parseResult,
             shebang,
@@ -869,9 +867,8 @@ module.exports = (function() {
                     const rule = ruleCreator.create ? ruleCreator.create(ruleContext)
                         : ruleCreator(ruleContext);
 
-                    // add all the node types as listeners
+                    // add all the selectors from the rule as listeners
                     Object.keys(rule).forEach(selector => {
-                        selectors.add(selector);
                         api.on(selector, timing.enabled
                             ? timing.time(key, rule[selector])
                             : rule[selector]
@@ -914,7 +911,7 @@ module.exports = (function() {
                 }
             }
 
-            let eventGenerator = new NodeEventGenerator(api, selectors);
+            let eventGenerator = new NodeEventGenerator(api);
 
             eventGenerator = new CodePathAnalyzer(eventGenerator);
             eventGenerator = new CommentEventGenerator(eventGenerator, sourceCode);