--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -5,6 +5,8 @@
 
 "use strict";
 
+const esquery = require("esquery");
+
 //------------------------------------------------------------------------------
 // Public Interface
 //------------------------------------------------------------------------------
@@ -24,10 +26,20 @@
 class NodeEventGenerator {
 
     /**
-     * @param {EventEmitter} emitter - An event emitter which is the destination of events.
-     */
-    constructor(emitter) {
+    * @param {EventEmitter} emitter - An event emitter which is the destination of events.
+    * @param {string[]} selectors A list of query selector strings to detect and emit
+    * @returns {NodeEventGenerator} new instance
+    */
+    constructor(emitter, selectors) {
         this.emitter = emitter;
+        this.currentAncestry = [];
+        this.parsedQueryMap = (selectors || [])
+
+            // Filter out selectors that only contain a node name, and treat them as a special case.
+            // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
+            // always safe to just emit a node's type while entering it.
+            .filter(query => !/^\w+(:exit)?$/.test(query))
+            .reduce((parsedQueryMap, selector) => parsedQueryMap.set(selector, esquery.parse(selector)), new Map());
     }
 
     /**
@@ -37,6 +49,16 @@ class NodeEventGenerator {
      */
     enterNode(node) {
         this.emitter.emit(node.type, node);
+
+        if (node.parent) {
+            this.currentAncestry.unshift(node.parent);
+        }
+
+        this.parsedQueryMap.forEach((parsedQuery, rawQuery) => {
+            if (esquery.matches(node, parsedQuery, this.currentAncestry)) {
+                this.emitter.emit(rawQuery, node);
+            }
+        });
     }
 
     /**
@@ -45,6 +67,7 @@ class NodeEventGenerator {
      * @returns {void}
      */
     leaveNode(node) {
+        this.currentAncestry.shift();
         this.emitter.emit(`${node.type}:exit`, node);
     }
 }