--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -7,12 +7,11 @@
 
 const esquery = require("esquery");
 const lodash = require("lodash");
-const scheduledExitMatches = new WeakMap();
 
 /**
 * Gets the possible types of a selector
 * @param {Object} selector parsed selector
-* @returns {string[]|null} return value (TODO)
+* @returns {string[]|null} The node types that could possibly trigger this selector, or `null` if all node types could trigger it
 */
 function getPossibleTypes(selector) {
     switch (selector.type) {
@@ -43,6 +42,15 @@ function getPossibleTypes(selector) {
     }
 }
 
+class Selector {
+    constructor(query) {
+        this.query = query;
+        this.parsedSelector = esquery.parse(query.replace(/:exit$/, ""));
+        this.isExit = query.endsWith(":exit");
+        this.listenerTypes = getPossibleTypes(this.parsedSelector);
+    }
+}
+
 //------------------------------------------------------------------------------
 // Public Interface
 //------------------------------------------------------------------------------
@@ -69,30 +77,42 @@ class NodeEventGenerator {
     constructor(emitter, rawSelectors) {
         this.emitter = emitter;
         this.currentAncestry = [];
-        this.parsedQueryMap = (rawSelectors || [])
+        this.scheduledExitMatches = new WeakMap();
+        this.selectors = (rawSelectors || [])
 
             // Filter out selectors that only contain a node name, and treat them as a special case.
             // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
             // always safe to just emit a node's type while entering it.
-            .filter(query => !/^\w+(:exit)?$/.test(query))
-            .reduce((parsedQueryMap, rawSelector) => parsedQueryMap.set(esquery.parse(rawSelector.replace(/:exit$/, "")), rawSelector), new Map());
+            .filter(query => !/(^\w+|\*)(:exit)?$/.test(query))
+            .map(query => new Selector(query));
 
         this.selectorsByNodeType = new Map();
         this.anyTypeSelectors = [];
-        this.parsedQueryMap.forEach((_, parsedSelector) => {
-            const possibleTypes = getPossibleTypes(parsedSelector);
-
-            if (possibleTypes) {
-                possibleTypes.forEach(nodeType => {
+        this.selectors.forEach(selector => {
+            if (selector.listenerTypes) {
+                selector.listenerTypes.forEach(nodeType => {
                     if (!this.selectorsByNodeType.has(nodeType)) {
                         this.selectorsByNodeType.set(nodeType, []);
                     }
-                    this.selectorsByNodeType.get(nodeType).push(parsedSelector);
+                    this.selectorsByNodeType.get(nodeType).push(selector);
                 });
             } else {
-                this.anyTypeSelectors.push(parsedSelector);
+                this.anyTypeSelectors.push(selector);
             }
         });
+    }
+
+    applySelector(node, selector) {
+        if (esquery.matches(node, selector.parsedSelector, this.currentAncestry)) {
+            if (selector.isExit) {
+                if (!this.scheduledExitMatches.has(node)) {
+                    this.scheduledExitMatches.set(node, []);
+                }
+                this.scheduledExitMatches.get(node).push(selector);
+            } else {
+                this.emitter.emit(selector.query, node);
+            }
+        }
     }
 
     /**
@@ -102,25 +122,17 @@ class NodeEventGenerator {
      */
     enterNode(node) {
         this.emitter.emit(node.type, node);
+        this.emitter.emit("*", node);
 
         if (node.parent) {
             this.currentAncestry.unshift(node.parent);
         }
 
-        const selectorsToAttempt = this.anyTypeSelectors.concat(this.selectorsByNodeType.get(node.type) || []);
+        this.anyTypeSelectors.forEach(selector => this.applySelector(node, selector));
 
-        selectorsToAttempt
-            .filter(selector => esquery.matches(node, selector, this.currentAncestry))
-            .forEach(selector => {
-                if (this.parsedQueryMap.get(selector).endsWith(":exit")) {
-                    if (!scheduledExitMatches.has(node)) {
-                        scheduledExitMatches.set(node, []);
-                    }
-                    scheduledExitMatches.get(node).push(selector);
-                } else {
-                    this.emitter.emit(this.parsedQueryMap.get(selector), node);
-                }
-            });
+        if (this.selectorsByNodeType.has(node.type)) {
+            this.selectorsByNodeType.get(node.type).forEach(selector => this.applySelector(node, selector));
+        }
     }
 
     /**
@@ -131,9 +143,10 @@ class NodeEventGenerator {
     leaveNode(node) {
         this.currentAncestry.shift();
         this.emitter.emit(`${node.type}:exit`, node);
+        this.emitter.emit("*:exit", node);
 
-        if (scheduledExitMatches.has(node)) {
-            scheduledExitMatches.get(node).forEach(selector => this.emitter.emit(selector, node));
+        if (this.scheduledExitMatches.has(node)) {
+            this.scheduledExitMatches.get(node).forEach(selector => this.emitter.emit(selector.parsedSelector, node));
         }
     }
 }