--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -42,14 +42,16 @@ function getPossibleTypes(selector) {
     }
 }
 
-class Selector {
-    constructor(query) {
-        this.query = query;
-        this.parsedSelector = esquery.parse(query.replace(/:exit$/, ""));
-        this.isExit = query.endsWith(":exit");
-        this.listenerTypes = getPossibleTypes(this.parsedSelector);
-    }
-}
+const parseSelector = lodash.memoize(query => {
+    const parsedSelector = esquery.parse(query.replace(/:exit$/, ""));
+
+    return {
+        query,
+        isExit: query.endsWith(":exit"),
+        parsedSelector,
+        listenerTypes: getPossibleTypes(parsedSelector)
+    };
+});
 
 //------------------------------------------------------------------------------
 // Public Interface
@@ -84,7 +86,7 @@ class NodeEventGenerator {
             // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
             // always safe to just emit a node's type while entering it.
             .filter(query => !/(^\w+|\*)(:exit)?$/.test(query))
-            .map(query => new Selector(query));
+            .map(parseSelector);
 
         this.selectorsByNodeType = new Map();
         this.anyTypeSelectors = [];