--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -73,23 +73,29 @@ class NodeEventGenerator {
 
     /**
     * @param {EventEmitter} emitter - An event emitter which is the destination of events.
-    * @param {string[]} rawSelectors A list of query selector strings to detect and emit
+    * @param {Set<string>} queries A list of query selector strings to detect and emit
     * @returns {NodeEventGenerator} new instance
     */
-    constructor(emitter, rawSelectors) {
+    constructor(emitter, queries) {
         this.emitter = emitter;
         this.currentAncestry = [];
         this.scheduledExitMatches = new WeakMap();
-        this.selectors = (rawSelectors || [])
-
-            // Filter out selectors that only contain a node name, and treat them as a special case.
-            // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
-            // always safe to just emit a node's type while entering it.
-            .filter(query => !/(^\w+|\*)(:exit)?$/.test(query))
-            .map(parseSelector);
-
+        this.selectors = [];
         this.selectorsByNodeType = new Map();
         this.anyTypeSelectors = [];
+
+        if (queries) {
+            queries.forEach(query => {
+
+                // Filter out selectors that only contain a node name, and treat them as a special case.
+                // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
+                // always safe to just emit a node's type while entering it.
+                if (!/(^\w+|\*)(:exit)?$/.test(query)) {
+                    this.selectors.push(parseSelector(query));
+                }
+            });
+        }
+
         this.selectors.forEach(selector => {
             if (selector.listenerTypes) {
                 selector.listenerTypes.forEach(nodeType => {