--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -90,7 +90,7 @@ class NodeEventGenerator {
                 // Filter out selectors that only contain a node name, and treat them as a special case.
                 // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
                 // always safe to just emit a node's type while entering it.
-                if (!/(^\w+|\*)(:exit)?$/.test(query)) {
+                if (!/^\w+(:exit)?$/.test(query)) {
                     this.selectors.push(parseSelector(query));
                 }
             });
@@ -130,7 +130,6 @@ class NodeEventGenerator {
      */
     enterNode(node) {
         this.emitter.emit(node.type, node);
-        this.emitter.emit("*", node);
 
         if (node.parent) {
             this.currentAncestry.unshift(node.parent);
@@ -151,7 +150,6 @@ class NodeEventGenerator {
     leaveNode(node) {
         this.currentAncestry.shift();
         this.emitter.emit(`${node.type}:exit`, node);
-        this.emitter.emit("*:exit", node);
 
         if (this.scheduledExitMatches.has(node)) {
             this.scheduledExitMatches.get(node).forEach(selector => this.emitter.emit(selector.parsedSelector, node));