--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -84,17 +84,15 @@ class NodeEventGenerator {
         this.selectorsByNodeType = new Map();
         this.anyTypeSelectors = [];
 
-        if (queries) {
-            queries.forEach(query => {
+        queries.forEach(query => {
 
-                // Filter out selectors that only contain a node name, and treat them as a special case.
-                // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
-                // always safe to just emit a node's type while entering it.
-                if (!/^\w+(:exit)?$/.test(query)) {
-                    this.selectors.push(parseSelector(query));
-                }
-            });
-        }
+            // Filter out selectors that only contain a node name, and treat them as a special case.
+            // Although these selectors are valid, it's not necessary to check all of them on every node, because it's
+            // always safe to just emit a node's type while entering it.
+            if (!/^\w+(:exit)?$/.test(query)) {
+                this.selectors.push(parseSelector(query));
+            }
+        });
 
         this.selectors.forEach(selector => {
             if (selector.listenerTypes) {