--- a/lib/util/node-event-generator.js
+++ b/lib/util/node-event-generator.js
@@ -19,16 +19,28 @@ function getPossibleTypes(selector) {
             return [selector.value];
 
         case "matches": {
-            const typesPerCase = selector.selectors.map(getPossibleTypes);
+            const typesForComponents = selector.selectors.map(getPossibleTypes);
 
-            if (typesPerCase.every(typeList => typeList)) {
-                return lodash.union.apply(null, typesPerCase);
+            if (typesForComponents.every(typesForComponent => typesForComponent)) {
+                return lodash.union.apply(null, typesForComponents);
             }
             return null;
         }
 
-        case "compound":
-            return lodash.intersection.apply(null, selector.selectors.map(getPossibleTypes).filter(Boolean));
+        case "compound": {
+            const typesForComponents = selector.selectors.map(getPossibleTypes).filter(typesForComponent => typesForComponent);
+
+            // If all of the components could match any type, then the compound could also match any type.
+            if (!typesForComponents.length) {
+                return null;
+            }
+
+            /*
+             * If at least one of the components could only match a particular type, the compound could only match
+             * the intersection of those types.
+             */
+            return lodash.intersection.apply(null, typesForComponents);
+        }
 
         case "child":
         case "descendant":
@@ -150,7 +162,7 @@ class NodeEventGenerator {
         this.emitter.emit(`${node.type}:exit`, node);
 
         if (this.scheduledExitMatches.has(node)) {
-            this.scheduledExitMatches.get(node).forEach(selector => this.emitter.emit(selector.parsedSelector, node));
+            this.scheduledExitMatches.get(node).forEach(selector => this.emitter.emit(selector.query, node));
         }
     }
 }