… docs/developer-guide/selectors.md
… docs/developer-guide/working-with-rules.md
… docs/rules/no-restricted-syntax.md
… package.json
… tests/lib/code-path-analysis/code-path-analyzer.js
… tests/lib/rules/no-restricted-syntax.js
… tests/lib/util/comment-event-generator.js
… tests/lib/util/node-event-generator.js