**What is the purpose of this pull request? (put an "X" next to item)**

[x] Add something to the core

**What changes did you make? (Give an overview)**

This updates node-event-generator to allow rules to listen for selectors. It also updates the implementation of a few simple rules to use this feature.

Before traversal begins, all the selectors needed by the active rules are gathered and passed to `node-event-generator`. When a node is entered, `node-event-generator` checks whether the node matches each selector using esquery, and emits the selector if so. Since this implementation detects shallow matches for a given node rather than getting all matches in the AST for a selector, it only needs to do one AST traversal.

~~This PR is not yet complete. Remaining things to do:~~

* ~~Decide whether we want to go forward with this approach~~
* ~~Call selector listeners in a consistent order. Currently, if a rule is listening for multiple selectors that match the same node, they fire in a different order depending on whether other rules also listen for those selectors. This is undesirable -- we should guarantee that listeners always fire in the same order.~~
* ~~Determine how to support `:exit` with selectors. (Currently, `Foo:exit` is valid like it was before, but something like `Foo > Bar:exit` is not.)~~
* ~~Add documentation~~
* ~~Add more tests~~
* ~~Make sure selector specificity definition is consistent with CSS~~
* ~~Fail gracefully if a rule specifies an invalid selector~~

Also see: https://github.com/eslint/eslint/pull/5441, https://github.com/eslint/eslint/pull/7822

**Is there anything you'd like reviewers to focus on?**

I'm interested in feedback on this approach to detecting selector matches.