@not-an-aardvark 2017-01-21
As discussed with @ilyavolodin, I'm working on a modified query matching solution optimized for our use-case (i.e. doing all matches in a single traversal)

@ilyavolodin This is far from complete (there are currently a bunch of tests failing), but could you double-check that the approach in `lib/util/node-event-generator.js` is roughly what you had in mind?