@ilyavolodin 2017-01-25
@not-an-aardvark This is along the lines I was thinking about, but I think it can still be improved. First of all, after you parsed selector, I would create a map with a list of nodeTypes specified in the selector, for example `CallExpression > MemberExpression.callee[object.name = 'process'][property.name = 'exit']` would be turned into a map with two elements `CallExpression` and `MemberExpression`. Then when `enterNode` is called with `CallExpression` node, you can mark it as visited in the map (and mark it as non-visited on `CallExpression:exit`). This way, when you hit `enterNode` with `MemberExpression` you can check your map to see if all nodes are marked as visited, and only then do evaluation of the selector.
There are a few issues to watch out for in this approach. First of all, if you hit two `CallExpression` before you hit `MemberExpression` you might hit `CallExpression:exit` before `MemberExpression` and that would clear visited flag. Second you should only mark something as visited if all of the nodes on the left of current node in the selector are already marked as visited. This approach also doesn't support sibling selectors (not sure how to handle those).
Also, if performance doesn't improve with this approach, you could run `\[a-zA-Z]\` regEx on the selector, and if it passes, you can just bypass all of the selector code.