@not-an-aardvark 2017-01-25
I'm concerned that approach might actually hurt performance more than it would help it, because it will have to do a lot of unnecessary checks. For example, consider the selector `ArrayExpression > CallExpression > Identifier`. If the traverser encounters a large array, it will have to check every single element of the array to see if it matches `CallExpression`. I'm not sure I see how this would be an improvement over the way that esquery does it.

I'm basing some of this off of [this explanation](https://stackoverflow.com/questions/5797014/why-do-browsers-match-css-selectors-from-right-to-left) about CSS matching.