@ilyavolodin 2017-01-25
Hmm... You might be right. I was trying to avoid going up-and-down the tree too many times. So in that case, your implementation is basically matching what @nzakas suggested (without additional emitter). I'm fine with that. I think we should check performance hit of this approach too. Previous implementation was not very fast.

For future reference, numbers with the first implementations were:

Single file: 3443.6669190000002ms -> 4689.529589ms.

Multifile tests: 12095.475817ms -> 13363.122209ms