@not-an-aardvark 2017-01-25
> So in that case, your implementation is basically matching what @nzakas suggested (without additional emitter)

To clarify, are you talking about the original implementation in https://github.com/eslint/eslint/commit/b9a69683df3498ae322d76caf4826d5fb066504b (which uses esquery), or the current implementation in https://github.com/eslint/eslint/commit/3c370e21501cd6370d2c8ffacf745a4b79df68a0 (which uses a custom parser/matcher without esquery)?

At this point I'm thinking it might be better to use esquery along with basic type prediction from esquery's parser. Aside from type prediction, I don't think I added anything in https://github.com/eslint/eslint/commit/3c370e21501cd6370d2c8ffacf745a4b79df68a0 that esquery doesn't already do.