@ilyavolodin 2017-01-25
I was talking about current implementation. Specifically on running selector matching only if the right-most selector matches the current node. I agree that we can try doing it with esquery, at least to get the perf numbers.