@not-an-aardvark 2017-01-26
https://github.com/eslint/eslint/pull/7833/commits/c802303378124eb24d20ee61abc834adc097e8ea uses esquery with a bit of additional optimization to filter selectors based on node types. This results in a 3% performance decrease for the single-file test and an 8% performance decrease for the multi-file test, which seems significantly better than before. It might be possible to optimize further.