@ilyavolodin 2017-01-26
3% is not too bad, I think that's something we can live with. But 8% on multiple files is not great. Any ideas for further optimizations? Slowdown on multiple files suggests that this has something to do with with constructor.