@not-an-aardvark 2017-03-09
I've improved performance a bit by memoizing selector-parsing calls.

The current performance penalties are:

* single-file: 4.4% decrease
* multi-file: 2.1% decrease