@ilyavolodin 2017-03-09
Nice. I think we could live with 4% bump for single files given the advantages of this approach. We can also always try to optimize this later as well.