@ilyavolodin 2017-03-09
Builds are failing because this PR pushes coverage below threshold.