@not-an-aardvark 2017-03-09
Yeah, I'm in the process of adding more tests now. Right now there's no coverage for a lot of the selector-type logic.