@not-an-aardvark 2017-03-09
@ilyavolodin Do you have any thoughts on the best way to do this?

> [TODO]: Call selector listeners in a consistent order. Currently, if a rule is listening for multiple selectors that match the same node, they fire in a different order depending on whether other rules also listen for those selectors. This is undesirable -- we should guarantee that listeners always fire in the same order.