@ilyavolodin 2017-03-09
@not-an-aardvark Could you clarify the comment about order of the events? Maybe a short example?